Remove custom template based on post id
I don't like template customization based on post id, I should be able
to swap out the slider and it should display in the same manner.

Therefore, I've removed the template override based on post id.

We now only have a general template override in the theme (or one
specific to the template attribute passed in via the shortcode, e.g.
[fe_slider id=5 template=footer]

See #2